refactor(omo-native): split embedded runtime provisioning out of compile-entry
