+++ packages/omo-native/compile-entry.ts
@@ -1,9 +1,19 @@
-import { createHash } from "node:crypto"
-import { chmodSync, copyFileSync, existsSync, mkdirSync, readFileSync, realpathSync, renameSync, rmSync, writeFileSync } from "node:fs"
+import { existsSync } from "node:fs"
 import { homedir } from "node:os"
-import { dirname, join, resolve } from "node:path"
+import { dirname, join } from "node:path"
 import { spawn } from "node:child_process"
 import { fileURLToPath } from "node:url"
+import {
+  embeddedText,
+  isProvisionedExecutable,
+  materializeProvisionedExecutable,
+  provisionEmbeddedRuntime,
+  runningExecutablePath,
+  selectRuntimeManifest,
+  shouldReexecAfterProvisioning,
+  type EmbeddedFile,
+  type EmbeddedManifest,
+} from "./compile-runtime"
 import { migrateLegacyBunGlobalManifest } from "./bin/lib/legacy-bun-global-migration.js"
 import { adoptLegacyFlatState, canonicalAgentDir } from "./bin/lib/agent-dir.js"
 import { nearestNodeBin, readJson } from "./bin/lib/package-paths.js"
@@ -11,15 +21,6 @@ import { runDoctor } from "./bin/lib/doctor.js"
 import { detectHarnesses, needsSetupSuggestion } from "./bin/lib/setup-detect.js"
 import { printSetupReport } from "./bin/lib/setup-report.js"
 import { delimiter } from "node:path"
-
-type EmbeddedFile = Blob & {
-  name: string
-  arrayBuffer?: () => Promise<ArrayBuffer>
-  bytes?: () => Promise<Uint8Array>
-  text?: () => Promise<string>
-}
-export type EmbeddedManifestEntry = { relPath: string; sha256: string; mode: number; size: number }
-export type EmbeddedManifest = { omoAiVersion: string; enginePin: string; manifestSha: string; entries: EmbeddedManifestEntry[] }
 
 // The engine is imported via a RELATIVE string LITERAL, inlined at both import
 // sites, and both properties are load-bearing:
@@ -65,48 +66,6 @@ export function updateLine(platform: NodeJS.Platform, arch: string): string {
   return `omo is updated via curl: curl -fsSL https://github.com/code-yeongyu/oh-my-openagent/releases/latest/download/${asset} -o ${dest} && chmod +x ${dest}`
 }
 
-export function isProvisionedExecutable(execPath: string, expectedPath: string): boolean {
-  try {
-    return realpathSync(execPath) === realpathSync(expectedPath)
-  } catch {
-    return resolve(execPath) === resolve(expectedPath)
-  }
-}
-
-export function materializeProvisionedExecutable(
-  sourcePath: string,
-  destinationPath: string,
-  platform = process.platform,
-): void {
-  if (platform === "win32") {
-    if (existsSync(destinationPath)) return
-    copyFileSync(sourcePath, destinationPath)
-    chmodSync(destinationPath, 0o755)
-    return
-  }
-  const temporaryPath = `${destinationPath}.tmp-${process.pid}`
-  try {
-    rmSync(temporaryPath, { force: true })
-    copyFileSync(sourcePath, temporaryPath)
-    chmodSync(temporaryPath, 0o755)
-    renameSync(temporaryPath, destinationPath)
-  } finally {
-    rmSync(temporaryPath, { force: true })
-  }
-}
-
-export function runningExecutablePath(
-  argv0 = process.argv[0],
-  execPath = process.execPath,
-  platform = process.platform,
-): string {
-  return platform === "win32" && argv0.toLowerCase().endsWith(".exe") ? argv0 : execPath
-}
-
-export function shouldReexecAfterProvisioning(platform = process.platform): boolean {
-  return platform !== "win32"
-}
-
 export function remapSenpiEnvironment(source: NodeJS.ProcessEnv = process.env, execDir: string): NodeJS.ProcessEnv {
   const env = { ...source }
   delete env.OMO_BIN
@@ -133,60 +92,6 @@ export function remapSenpiEnvironment(source: NodeJS.ProcessEnv = process.env, e
   }
   env.OMO_BIN = join(execDir, process.platform === "win32" ? "omo.exe" : "omo")
   return env
-}
-
-async function embeddedText(file: EmbeddedFile): Promise<string> {
-  if (file.text) return file.text()
-  if (file.arrayBuffer) return Buffer.from(await file.arrayBuffer()).toString("utf8")
-  throw new Error(`embedded asset ${file.name} cannot be read`)
-}
-
-async function embeddedBytes(file: EmbeddedFile): Promise<Uint8Array> {
-  if (file.bytes) return file.bytes()
-  if (file.arrayBuffer) return new Uint8Array(await file.arrayBuffer())
-  throw new Error(`embedded asset ${file.name} cannot be read as bytes`)
-}
-
-export async function selectRuntimeManifest(embedded: EmbeddedFile[]): Promise<EmbeddedFile | undefined> {
-  const exact = embedded.find((file) => file.name === "omo-runtime/runtime-manifest.json")
-  if (exact) return exact
-  for (const file of embedded) {
-    if (!file.name.endsWith("runtime-manifest.json")) continue
-    try {
-      const parsed = JSON.parse(await embeddedText(file))
-      if (typeof parsed?.omoAiVersion === "string" && typeof parsed?.enginePin === "string") return file
-    } catch {
-      // Non-manifest assets with a similar name are not candidates.
-    }
-  }
-  return undefined
-}
-
-export async function provisionEmbeddedRuntime(manifest: EmbeddedManifest, embedded: EmbeddedFile[], runtimeDir: string): Promise<void> {
-  mkdirSync(runtimeDir, { recursive: true })
-  const marker = join(runtimeDir, ".provisioned")
-  if (readFileIfExists(marker)?.trim() === manifest.manifestSha) return
-  const byPath = new Map(embedded.map((file) => [
-    file.name.replace(/^\.\//, "").replace(/^omo-runtime\//, ""),
-    file,
-  ]))
-  for (const entry of manifest.entries) {
-    const file = byPath.get(entry.relPath.replace(/^\.\//, ""))
-    if (!file) throw new Error(`embedded asset missing: ${entry.relPath}`)
-    const bytes = await embeddedBytes(file)
-    if (bytes.byteLength !== entry.size || createHash("sha256").update(bytes).digest("hex") !== entry.sha256) {
-      throw new Error(`embedded asset integrity mismatch: ${entry.relPath}`)
-    }
-    const destination = join(runtimeDir, entry.relPath)
-    mkdirSync(dirname(destination), { recursive: true })
-    writeFileSync(destination, bytes, { mode: entry.mode })
-    chmodSync(destination, entry.mode)
-  }
-  writeFileSync(marker, `${manifest.manifestSha}\n`, { mode: 0o644 })
-}
-
-function readFileIfExists(path: string): string | undefined {
-  try { return readFileSync(path, "utf8") } catch { return undefined }
 }
 
 function runCompiledDoctor(inventory: Awaited<ReturnType<typeof detectHarnesses>>, execDir: string, enginePin: string): void {
